Split joystickd into joystickd and joystick_control
